feat: Add interactivity to divs and spans
This commit adds JavaScript functionality to change the color of divs and spans when interacting with them. It also adds event listeners to handle mouseenter, mouseleave, click, and dblclick events. Additionally, it introduces the ability to display an alert message when clicking on a div with both "c1" and "c2" classes.